Reversed :float instruction argument order where operations are not commutative.

+++ instructions/float/divide.rb
@@ -2,8 +2,8 @@ class FloatDivide < NudgeInstruction
   get 2, :float
   
   def process
-    if float(1) != 0
-      put :float, float(0) / float(1)
+    if float(0) != 0
+      put :float, float(1) / float(0)
     else
       raise NudgeError::DivisionByZero, "cannot divide a float by 0"
     end
