+++ instructions/float/subtract.rb
@@ -2,6 +2,6 @@ class FloatSubtract < NudgeInstruction
   get 2, :float
   
   def process
-    put :float, float(0) - float(1)
+    put :float, float(1) - float(0)
   end
 end
